(fix): Se corrigió en el módulo cotizaciones: obtener el usuario_id apartir del token para luego registrar una cotizacion
